Refactor largest .rs files to adhere to SRP
- Identified `coercion.rs` and `system_procedures.rs` as the largest `iridium_core` files violating SRP.
- Split `coercion.rs` into `numeric.rs`, `string.rs`, `datetime.rs`, and `binary.rs` submodules.
- Split `system_procedures.rs` into `metadata.rs`, `session.rs`, and `security.rs` submodules.
- Maintained all existing functionality and addressed local visibility/compilation bounds.
