+++ crates/iridium_core/src/executor/value_ops/mod.rs
@@ -3,6 +3,6 @@ pub mod comparison;
 pub mod formatting;
 
 pub use coercion::coerce_value_to_type_with_dateformat;
-pub use coercion::parse_numeric_literal;
+pub use coercion::numeric::parse_numeric_literal;
 pub use comparison::{compare_values, truthy};
 pub use formatting::convert_with_style;
